Move TypeScript files into ts/ folder
