+++ tsconfig.json
@@ -11,7 +11,7 @@
     "paths": {
       "lexicon/*": ["lexicon/static/jsx/search/*"],
       "narratives/*": ["narratives/static/jsx/*"],
-      "query-builder/*": ["query_builder/static/jsx/*"],
+      "query-builder/*": ["query_builder/static/ts/*"],
       "plyr-src/*": ["node_modules/plyr/src/*"],
     }
   }
