+++ webpack.config.js
@@ -67,7 +67,7 @@ module.exports = {
     alias: {
       lexicon: path.resolve(__dirname, 'lexicon/static/jsx/search/'),
       narratives: path.resolve(__dirname, 'narratives/static/jsx/'),
-      'query-builder': path.resolve(__dirname, 'query_builder/static/jsx/'),
+      'query-builder': path.resolve(__dirname, 'query_builder/static/ts/'),
       'plyr-src': path.resolve(__dirname, 'node_modules/plyr/src/'),
     },
   },
